+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Unreleased
+
+# v0.1.0-rc4
 -  [#908](https://github.com/kubernetes-sigs/kubefed/issues/908) Adds admission
    webhook validations for KubeFedCluster API.
 -  [#982](https://github.com/kubernetes-sigs/kubefed/issues/982) To